[update]ヒール量、攻撃力をint

+++ Unity1Week2026.03/Assets/People/Miyamoto/Script/Effects/AttackBase.cs
@@ -3,6 +3,6 @@
 public abstract class AttackBase : ICardEffect
 {
     [SerializeField]
-    protected float _attackPower;
+    protected int _attackPower;
     public abstract void Exucute();
 }
