+++ Unity1Week2026.03/Assets/People/Miyamoto/Script/Effects/HealBase.cs
@@ -3,6 +3,6 @@
 public abstract class HealBase : ICardEffect
 {
     [SerializeField]
-    protected float _healAmount;
+    protected int _healAmount;
     public abstract void Exucute();
 }
